server/requirements.txt
pandas
numpy
scikit-learn
mlflow
matplotlib
flask
flask-cors
requests
ipykernel
urllib3
google-generativeai
dotenv